refactor(#13): extract CLI command orchestration into modular builders and runners
Split monolithic src/commands/new.ts (1093 lines) into:
- src/types.ts: interfaces and path constants
- src/utils.ts: validation, JWT generation, utilities
- src/prompts/project.prompts.ts: user input prompts
- src/builders/{api,web,db}.builder.ts: project scaffolding
- src/runners/{terminal,server,database}.runner.ts: lifecycle management

Improvements:
- Reduced new.ts from 1093 to 253 lines (pure orchestration)
- Improved maintainability through separation of concerns
- Fixed regex regression in email service template generation
- Added missing signup error translation keys
