+++ src/types.ts
@@ -1,0 +1,60 @@
+import { resolve } from 'path'
+
+export interface DbCredentials {
+  host: string
+  port: string
+  user: string
+  password: string
+  database: string
+  dbType: 'postgresql' | 'sql'
+}
+
+export interface Answers {
+  projectName: string
+  projectDescription: string
+  isMonorepo: boolean
+  setupRepo: 'local' | 'create' | 'existing'
+  gitProvider?: 'GitHub' | 'GitLab'
+  mainBranch: 'main' | 'master'
+  monorepoUrl?: string
+  backendRepoUrl: string
+  frontendRepoUrl?: string
+  dbSetup: 'docker' | 'credentials' | 'manual'
+  dbCredentials?: DbCredentials
+  initDb: boolean
+  emailService: 'none' | 'mailersend'
+  mailersendApiKey?: string
+  mailersendSenderEmail?: string
+  mailersendSenderName?: string
+}
+
+export interface CreateApiAppParams {
+  isMonorepo: boolean
+  projectName: string
+  projectDescription: string
+  backendRepoUrl: string
+  dbCredentials?: DbCredentials
+  mainBranch: string
+  emailService: 'none' | 'mailersend'
+  mailersendApiKey?: string
+  mailersendSenderEmail?: string
+  mailersendSenderName?: string
+}
+
+export interface CreateWebAppParams {
+  isMonorepo: boolean
+  projectName: string
+  projectDescription: string
+  frontendRepoUrl: string
+  mainBranch: string
+}
+
+export interface CreateDbAppParams {
+  isMonorepo: boolean
+  projectName: string
+  dbCredentials?: DbCredentials
+}
+
+// Paths
+export const blueprintsPath = resolve(__dirname, '../scaffolds/blueprints')
+export const overlaysPath = resolve(__dirname, '../scaffolds/overlays')
