+++ src/utils.ts
@@ -1,0 +1,53 @@
+import crypto from 'crypto'
+import fs from 'fs'
+
+import { DbCredentials } from './types'
+
+/**
+ * Validate that a project name is safe for use in shell commands and file paths
+ */
+export function validateProjectName(name: string): void {
+  if (!/^[a-z0-9-]+$/.test(name)) {
+    throw new Error(`Invalid project name "${name}": can only contain lowercase letters, numbers, and hyphens`)
+  }
+}
+
+/**
+ * Generate a secure random string for JWT secrets
+ * @param length Length of the secret (default: 64)
+ * @returns A secure random string
+ */
+export function generateJwtSecret(length: number = 64): string {
+  return crypto.randomBytes(length).toString('hex')
+}
+
+/**
+ * Set default values for database credentials if they are empty
+ */
+export function setDefaultDbCredentials(credentials?: DbCredentials): DbCredentials | undefined {
+  if (!credentials) return undefined
+
+  // define db type first
+  const dbType = credentials.dbType || 'postgresql'
+
+  return {
+    dbType,
+    host: credentials.host || 'localhost',
+    port: credentials.port || (dbType === 'postgresql' ? '5435' : '1433'),
+    user: credentials.user || 'db_dev_user',
+    password: credentials.password || 'db_dev_password',
+    database: credentials.database || 'db_dev'
+  }
+}
+
+/**
+ * Check if a file exists
+ */
+export async function fileExists(filePath: string): Promise<boolean> {
+  try {
+    await fs.promises.access(filePath, fs.constants.F_OK)
+    return true
+  } catch {
+    return false
+  }
+}
